Consume built Ceph binaries
Ceph is a large piece of software that take hours and hours to
build. It would be beneficial to the project and number of
iterations we can do per day if we can use the binaries directly.

Adjust snap definition to enable the binary and Python components
required to run a MON.

Add provisional `microceph` CLI written in BASH for repeatable
initialization steps during testing and development.

Add functional smoke test.