Build ceph-volume separately to work around hard coded user
Install snap with devmode, drop use of raw block device as
support is not really implemented and LVM is required for
operation of OSD and in line with future planned
disk configuration dependencies.

Add only one OSD until we figure out how to properly have
different backing stores for the `scsi_debug` kernel module or
find an alternative approach for testing with block devices.