ci: Move back to ubuntu-18.04 and use scsi_debug
To validate that preparing and adding raw block devices through
ceph-volmue works we need a "real" block device and not a loop
device.

We also most likely will not get an official snap interface with
raw access to /dev/loop* so testing with that for confined mode
is kind of pointless.