patch/ceph-volume: Don't chown if target has expected ownership
The Ceph processes will be running under a under a Mandatory
Access Control system as enforced by strict snap confinement.
Dropping privileges in such an environment is not necessary and
consequently performing the chown operation is not necessary
either.

We do not want to allow the snap to change ownership of arbritrary
special device files so we provide this patch instead to stop it
from trying to do it in the first place.