patch/ceph-volume: Improve patch to also handle gid
Check if gid matches as well as uid before calling `chown`, add an
extra check for target being owned by the 'disk' group.

Note that the goal is to not have to call out to chown at all when
not necessary as the `chown` call is not allowed when running in a
Mandatory Access Control confined environment.