Doc: need to restart microceph.daemon after dm-crypt conn
This is a stop gap measure for #190. There is an issue to track this
at https://bugs.launchpad.net/snapd/+bug/2034585 . For a more
permanent workaround we could alternatively also restart
microceph.daemon in an interface hook after dm-crypt connection,
however it'd be preferable to fix this in snapd so as to avoid API
interruptions.

Signed-off-by: Peter Sabaini <peter.sabaini@canonical.com>