microceph/cmd/microcephd: Add PreRemove hook to the microcluster daemon
Instead of directly calling the ceph service cleanup functions in the
CLI, this uses microcluster's `PreRemove` hook to clean up the services
before removing the node from the cluster.

To that end, the `removeNode` function has been wrapped in `PreRemove`
and it returns a function with a signature that can be accepted by
microcluster.Start as a hook.

The `node` field has been replaced with `state.Name` because this
function will only run on the node that is being removed, so its name
will be recorded in the State struct.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>