microceph/api: Use context.Background for async operations
microcluster no longer exposes the shutdown context, so these async
operations can no longer use it. It is not possible to use the request
context because it may cancel before the async operation returns, so
that just leaves context.Background.

In the future, MicroCeph should expose its own shutdown context that it
passes to microcluster app.Start.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>