Use LTS microcluster (#395)
Updates MicroCeph to use the stable release of microcluster at `v2.0.0`.

For the most part, the changes to the code in MicroCeph are just package
imports and changing the definition of `state.State` to an interface.

The actual behavioural changes are these:
* `state.Context` is no longer exposed, so `request.Context` from the
API handler is passed around explicitly to each helper. This should
actually result in more precise context handling for MicroCeph.
* One caveat is a few async API operations which spawn a goroutine. For
these `context.Background()` is used because the request context may
cancel before the goroutine is complete.

* Adds a schema update to MicroCeph that changes their foreign key
references to `core_cluster_members`. This table has been renamed from
`internal_cluster_members` to make it clear that accessing it is
expected behaviour. I also took the liberty of renaming the previous
schema update function to use an ordered naming scheme so that it
doesn't become ambiguous where an update fits among the others.

* Join tokens now have expiration dates. By default, MicroCeph will set
a 3 hour expiration for join tokens, but this can be configured with the
`--timeout` flag added to `cluster add`.

The project version supplied to microcluster has been set to `UNKOWN`
for now because microcluster requires a value here.

---------

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>