Add MicroCeph tutorial (#478)
# Description

We currently do not have a tutorial for new users in the MicroCeph
documentation. The how-to section of our documentation would be useful
for a user who is already familiar with Ceph and is looking to achieve
certain goals with MicroCeph. But, a new user with no prior knowledge of
Ceph would find it difficult to get started with MicroCeph, which is
intended to make their experience as easy as possible.

This is, hopefully, a complete, end-to-end tutorial intended to serve as
an entry point into Ceph for a new user, setting them up for their next
steps, whether it is to use MicroCeph to accomplish certain tasks in
their work (how-to guides), learn more about MicroCeph (explanation) or
make quick references.


## Type of change

- [x] Documentation update (Doc only change)

---------

Signed-off-by: Sharon Koech <sharon.koech@canonical.com>