Introduce maintenance mode (#482)
# Description

MicroCeph currently supports scaling up and down, but not maintenance.
In the actual operation, the operators may also need to perform regular
maintenance to the storage nodes, and MicroCeph does not offer a simple
solution to this problem. We would like to introduce sub-commands (e.g.
microceph cluster maintenance enter and microceph cluster maintenance
exit) to MicroCeph to enable the feature.

## Type of change

Please delete options that are not relevant.

- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] CleanCode (Code refactor, test updates, does not introduce
functional changes)
- [ ] Documentation update (Doc only change)

## How Has This Been Tested?

> **_NOTE:_** All functional changes should accompany corresponding
tests (unit tests, functional tests etc).

Tested in functional tests, and also manual test with 4 microceph nodes.

## Contributor's Checklist

Please check that you have:

- [x] self-reviewed the code in this PR.
- [x] added code comments, particularly in hard-to-understand areas.
- [ ] updated the user documentation with corresponding changes. (will
do in separated PR)
- [x] added tests to verify effectiveness of this change.

---------

Signed-off-by: Chi Wai Chan <chiwai.chan@canonical.com>