fix: improve pristine disk check with ceph-bluestore-tool validation (#621)
# Description

Improve the heuristics on what we consider a "pristine" disk. We used to
just check if the first 2k of a blockdev were zeroes, but with recent
upstream that isn't enough anymore. This change introduces more checks
so we're more strict on what we accept as pristine: checking for zero
sectors in various strategic locations (begin and end of disk, gpt
tables, 64k and 1M) and check if ceph-bluestore-tool returns any ceph
labels

If it fixes a reported issue, mention it here, i.e.:
Fixes #618 

## Type of change

Delete options that are not relevant.

- Bug fix (non-breaking change which fixes an issue)

## How has this been tested?

Add unit and functional test

## Contributor checklist

Please check that you have:

- [x] self-reviewed the code in this PR
- [x] added code comments, particularly in less straightforward areas
- [x] checked and added or updated relevant documentation
- [ ] checked and added or updated relevant release notes
- [x] added tests to verify effectiveness of this change

---------

Signed-off-by: Peter Sabaini <peter.sabaini@canonical.com>